Debug

+++ differentiable_robot_model/robot_model.py
@@ -778,36 +778,36 @@ class DifferentiableRobotModel(torch.nn.Module):
 
 
 class DifferentiableKUKAiiwa(DifferentiableRobotModel):
-    def __init__(self, device=None):
+    def __init__(self, device=None, dtype=torch.float32):
         rel_urdf_path = "kuka_iiwa/urdf/iiwa7.urdf"
         self.urdf_path = os.path.join(robot_description_folder, rel_urdf_path)
         self.learnable_rigid_body_config = None
         self.name = "differentiable_kuka_iiwa"
-        super().__init__(self.urdf_path, self.name, device=device)
+        super().__init__(self.urdf_path, self.name, device=device, dtype=dtype)
 
 
 class DifferentiableFrankaPanda(DifferentiableRobotModel):
-    def __init__(self, device=None):
+    def __init__(self, device=None, dtype=torch.float32):
         rel_urdf_path = "panda_description/urdf/panda_no_gripper.urdf"
         self.urdf_path = os.path.join(robot_description_folder, rel_urdf_path)
         self.learnable_rigid_body_config = None
         self.name = "differentiable_franka_panda"
-        super().__init__(self.urdf_path, self.name, device=device)
+        super().__init__(self.urdf_path, self.name, device=device, dtype=dtype)
 
 
 class DifferentiableTwoLinkRobot(DifferentiableRobotModel):
-    def __init__(self, device=None):
+    def __init__(self, device=None, dtype=torch.float32):
         rel_urdf_path = "2link_robot.urdf"
         self.urdf_path = os.path.join(robot_description_folder, rel_urdf_path)
         self.learnable_rigid_body_config = None
         self.name = "diff_2d_robot"
-        super().__init__(self.urdf_path, self.name, device=device)
+        super().__init__(self.urdf_path, self.name, device=device, dtype=dtype)
 
 
 class DifferentiableTrifingerEdu(DifferentiableRobotModel):
-    def __init__(self, device=None):
+    def __init__(self, device=None, dtype=torch.float32):
         rel_urdf_path = "trifinger_edu_description/trifinger_edu.urdf"
         self.urdf_path = os.path.join(robot_description_folder, rel_urdf_path)
         self.learnable_rigid_body_config = None
         self.name = "trifinger_edu"
-        super().__init__(self.urdf_path, self.name, device=device)
+        super().__init__(self.urdf_path, self.name, device=device, dtype=dtype)
